requirements.txt
PyYAML==6.0
pyudev==0.24.0